Cell altitudes and actor changes implemented
Cells now have an altitude which is set individually by the sublclasses

Significant actor changes, they now have a several new lists that tie in with a method to find which cells they would theoretically be able to move in. Additionally I've changed every actor subclass to use a super in its constructor to remove unessecary code.